requirements.txt
tflite-runtime==2.14.0
flask==3.1.0
flask-cors==5.0.0
numpy==1.26.4
pillow==11.0.0


